Commence work on the Recipes project through TOP
In this commit, I got carried away with setting up the content for the
recipes project for The Odin Project and forgot to make regular
commits. This commit encapsulates the core content of the recipes
including an index.html and four recipe pages in html.

Each page is linked to the index.html and vice versa. I will be tweaking
the project a little more before I move onto the next lesson.